requirements.txt
django
djangorestframework
requests
django-extensions
psycopg2-binary
django-filter